# core-project/docs/swagger.yaml
basePath: /
definitions:
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.Port:
    properties:
      code:
        type: string
      country:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  models.Voyage:
    properties:
      created_at:
        type: string
      destination_port_id:
        type: integer
      destinationPort:
        $ref: '#/definitions/models.Port'
      eta:
        type: string
      etd:
        type: string
      id:
        type: integer
      origin_port_id:
        type: integer
      originPort:
        $ref: '#/definitions/models.Port'
      updated_at:
        type: string
      vessel_name:
        type: string
      voyage_number:
        type: string
    type: object
host: localhost:8888
info:
  contact:
    email: support@container.local
    name: API Support
  description: REST API untuk manajemen container dengan autentikasi JWT.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Container Management API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password to receive a JWT token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login success, returns token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /api/auth/me:
    get:
      description: Returns the currently authenticated user's info extracted from
        JWT claims.
      produces:
      - application/json
      responses:
        "200":
          description: Current user info
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password. Password must
        be at least 6 characters.
      parameters:
      - description: User registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email is already registered
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
...
  /api/voyages:
    get:
      description: Get a list of all voyages, including origin and destination ports.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Voyage'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all voyages
      tags:
      - Voyages
    post:
      consumes:
      - application/json
      description: Create a new voyage. The `voyage_number` field must be unique.
      parameters:
      - description: Voyage payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Voyage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Voyage'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Voyage number already exists
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create voyage
      tags:
      - Voyages
  /api/voyages/{id}:
    delete:
      description: Delete a voyage by ID.
      parameters:
      - description: Voyage ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Voyage deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete voyage
      tags:
      - Voyages
    get:
      description: Retrieve a single voyage by its ID, including origin and destination
        ports.
      parameters:
      - description: Voyage ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Voyage'
        "404":
          description: Voyage not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get voyage by ID
      tags:
      - Voyages
    put:
      consumes:
      - application/json
      description: Update an existing voyage by ID.
      parameters:
      - description: Voyage ID
        in: path
        name: id
        required: true
        type: integer
      - description: Voyage payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Voyage'
      produces:
      - application/json
      responses:
        "200":
          description: Voyage updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Voyage number already exists
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update voyage
      tags:
      - Voyages
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"